Fixed a bug that concatinated rails together instead of ctarting new.
Added an option for Lumberjacking or Mining to prepare for the Recall-n-Rails Mining system.

Added a "loading rail" and "loaded rail" to let the user know that the rail has been selected correctly.

Added the starting location of the rail at the top of the file for future reference.

Changed the rail naming to reflect the starting location of the rail (example 2654-3256.txt) instead of rail.txt